Correciones de estilo de codigo y uso de clausulas Guard

+++ src/main/java/com/P1/Proyecto1Ruben_back/provider/Impl/ClienteProviderImpl.java
@@ -15,17 +15,18 @@ import com.P1.Proyecto1Ruben_back.repository.ClienteRepository;
 import jakarta.persistence.EntityNotFoundException;
 
 @Service
-public class ClienteProviderImpl implements ClienteProvider{
-	
+public class ClienteProviderImpl implements ClienteProvider {
+
 	@Autowired
 	private ClienteRepository repository;
-	
+
 	@Autowired
 	ModelMapper modelMapper;
 
 	@Override
 	public ClienteDto findClienteById(Long id) {
-		ClienteEntity clienteEntity = repository.findById(id).orElseThrow( () -> new EntityNotFoundException("El cliente con ID " + id + " no existe."));
+		ClienteEntity clienteEntity = repository.findById(id)
+				.orElseThrow(() -> new EntityNotFoundException("El cliente con ID " + id + " no existe."));
 		ClienteDto clienteDto = modelMapper.map(clienteEntity, ClienteDto.class);
 		return clienteDto;
 	}
@@ -38,66 +39,54 @@ public class ClienteProviderImpl implements ClienteProvider{
 	}
 
 	@Override
-	public ClienteDto updateCliente(Long id, ClienteDto cliente) {
-		ClienteEntity clienteEntity = repository.findById(id).orElseThrow( () -> new EntityNotFoundException("El cliente con ID " + id + " no existe."));
-		ClienteEntity clienteActualizado = modelMapper.map(cliente, ClienteEntity.class);
-		clienteEntity.setNombre(clienteActualizado.getNombre());
-		clienteEntity.setApellido1(clienteActualizado.getApellido1());
-		clienteEntity.setApellido2(clienteActualizado.getApellido2());
-		clienteEntity.setDireccion(clienteActualizado.getDireccion());
-		clienteEntity.setTfno(clienteActualizado.getTfno());
+	public ClienteDto updateCliente(ClienteDto cliente) {
+		ClienteEntity clienteEntity = repository.findById(cliente.getId())
+				.orElseThrow(() -> new EntityNotFoundException("El cliente con ID " + cliente.getId() + " no existe."));
+		modelMapper.map(cliente, clienteEntity);
 		repository.save(clienteEntity);
-		return cliente = modelMapper.map(clienteEntity, ClienteDto.class);
+		return cliente;
 	}
 
 	@Override
 	public void deleteClienteById(Long id) {
-		repository.findById(id).orElseThrow( () -> new EntityNotFoundException("El cliente con ID " + id + " no existe."));
+		if(id==null) {
+			throw new IllegalArgumentException ("El id no puede ser nulo.");
+		}
 		repository.deleteById(id);
 	}
 
 	@Override
 	public List<ClienteDto> allClients() {
 		List<ClienteEntity> listaEntity = repository.findAll();
-		List<ClienteDto> listaDto = new ArrayList<ClienteDto>();
-		for(ClienteEntity l :listaEntity) {
-			ClienteDto clienteDto = modelMapper.map(l, ClienteDto.class);
-			listaDto.add(clienteDto);
-		}
-		return listaDto;
+		return listaEntityAlistaDto(listaEntity);
 	}
 
 	@Override
 	public List<ClienteDto> findClienteByName(String name) {
-		if(!repository.findByName(name).isEmpty()){
-			List<ClienteEntity> listclienteEntity= repository.findByName(name);
-			List<ClienteDto> listaDto = new ArrayList<ClienteDto>();
-			for(ClienteEntity l :listclienteEntity) {
-				ClienteDto clienteDto = modelMapper.map(l, ClienteDto.class);
-				listaDto.add(clienteDto);
-			}
-			return listaDto;
-		}else {
+		if (repository.findByName(name).isEmpty()) {
 			throw new EntityNotFoundException("El cliente con nombre " + name + " no existe.");
 		}
-		
+		List<ClienteEntity> listclienteEntity = repository.findByName(name);
+		return listaEntityAlistaDto(listclienteEntity);
 	}
 
 	@Override
 	public List<ClienteDto> buscarClientes(String nombre, String apellido1, String apellido2, String direccion,
 			String tfno) {
-		if(!repository.findByNombreAndApellido1AndApellido2AndDireccionAndTfno(nombre, apellido1, apellido2, direccion, tfno).isEmpty()){
-			List<ClienteEntity> listclienteEntity= repository.findByNombreAndApellido1AndApellido2AndDireccionAndTfno(nombre, apellido1, apellido2, direccion, tfno);
-			List<ClienteDto> listaDto = new ArrayList<ClienteDto>();
-			for(ClienteEntity l :listclienteEntity) {
-				ClienteDto clienteDto = modelMapper.map(l, ClienteDto.class);
-				listaDto.add(clienteDto);
-			}
-			return listaDto;
-		}else {
+		if (repository.findByNombreAndApellido1AndApellido2AndDireccionAndTfno(nombre, apellido1, apellido2, direccion, tfno).isEmpty()) {
 			throw new EntityNotFoundException("El cliente con esos datos no existe.");
 		}
+		List<ClienteEntity> listclienteEntity = repository.findByNombreAndApellido1AndApellido2AndDireccionAndTfno(nombre, apellido1, 
+				apellido2, direccion, tfno);
+		return listaEntityAlistaDto(listclienteEntity);
 	}
 	
-
+	public List<ClienteDto> listaEntityAlistaDto(List<ClienteEntity> listaEntity) {
+		List<ClienteDto> listaDto = new ArrayList<ClienteDto>();
+		for (ClienteEntity l : listaEntity) {
+			ClienteDto clienteDto = modelMapper.map(l, ClienteDto.class);
+			listaDto.add(clienteDto);
+		}
+		return listaDto;
+	}
 }
